Extract to separate POSSendReceiptView
