+++ WooCommerce/Classes/POS/Presentation/CardReaderConnection/UI States/Reader Messages/PointOfSaleCardPresentPaymentSuccessMessageView.swift
@@ -8,11 +8,10 @@ struct PointOfSaleCardPresentPaymentSuccessMessageView: View {
 
     @State private var isShowingSendReceiptView: Bool = false
     @State private var isShowingReceiptNotEligibleBanner: Bool = false
-    @State private var textFieldInput: String = ""
 
     var body: some View {
         if isShowingSendReceiptView {
-            sendReceiptView
+            POSSendReceiptView(isShowingSendReceiptView: $isShowingSendReceiptView)
         } else {
             ZStack {
                 VStack(alignment: .center, spacing: Constants.headerSpacing) {
@@ -52,55 +51,6 @@ struct PointOfSaleCardPresentPaymentSuccessMessageView: View {
         }
     }
 
-    private var sendReceiptView: some View {
-        VStack(alignment: .leading, spacing: 20) {
-            HStack {
-                Button(action: {
-                    isShowingSendReceiptView = false
-                }, label: {
-                    Image(systemName: "arrow.backward")
-                        .font(.headline)
-                        .foregroundColor(.primary)
-                })
-                Spacer()
-            }
-            .padding()
-
-            Text(Localization.title)
-                .font(.largeTitle)
-                .bold()
-
-            Text(Localization.subtitle)
-                .font(.headline)
-
-            TextField(Localization.textfieldPlaceholder, text: $textFieldInput)
-                .keyboardType(.emailAddress)
-                .textInputAutocapitalization(.none)
-                .autocorrectionDisabled()
-                .textFieldStyle(RoundedBorderTextFieldStyle(focused: true))
-                .padding(.horizontal)
-
-            Button(action: {
-                Task { @MainActor in
-                    // TODO: Access posmodel
-                    // await posModel.sendReceipt(to: textFieldInput)
-                }
-            }, label: {
-                HStack(spacing: Constants.buttonSpacing) {
-                    Text(Localization.buttonTitle)
-                        .font(Constants.buttonFont)
-                }
-            })
-            .padding(Constants.buttonPadding)
-            .foregroundColor(Color.posPrimaryTextInverted)
-            .background(Color.posOverlayFillInverted)
-            .cornerRadius(Constants.buttonCornerRadius)
-
-            Spacer()
-        }
-        .padding()
-    }
-
     private var successIcon: some View {
         ZStack {
             Circle()
@@ -135,31 +85,6 @@ private extension PointOfSaleCardPresentPaymentSuccessMessageView {
         static let shadowSize: CGSize = .init(width: 0, height: 8)
         static let headerSpacing: CGFloat = 56
         static let textSpacing: CGFloat = 16
-        static let buttonSpacing: CGFloat = 12
-        static let buttonPadding: CGFloat = 32
-        static let buttonFont: POSFontStyle = .posBodyEmphasized
-        static let buttonCornerRadius: CGFloat = 8
-    }
-}
-
-private extension PointOfSaleCardPresentPaymentSuccessMessageView {
-    struct Localization {
-        static let title = NSLocalizedString(
-            "pointOfSale.sendreceipt.modal.title",
-            value: "Receipt",
-            comment: "Button title for the receipt button")
-        static let subtitle = NSLocalizedString(
-            "pointOfSale.sendreceipt.modal.subtitle",
-            value: "Email",
-            comment: "Subtitle for the view where an email address should be entered when sending receipts")
-        static let buttonTitle = NSLocalizedString(
-            "pointOfSale.sendreceipt.modal.button.title",
-            value: "Send",
-            comment: "Button title for sending a receipt")
-        static let textfieldPlaceholder = NSLocalizedString(
-            "pointOfSale.sendreceipt.modal.textfield.placeholder",
-            value: "Enter an email",
-            comment: "Placeholder for the view where an email address should be entered when sending receipts")
     }
 }
